Ports and updates
-Added "Back" button to loadstate screen, rearranged existing buttons
-Dupemod: Put resettting motion into it's own method, bc I was about to add `KeyBinding.unpressAllKeys()` to it
-AIManipulation|Forge: Extended the reach of  to 64 instead of 32
-AIManip+SpawnManip|Forge: Fixed box positions being misalinged with blocks
-DropManip|Forge: Scaled ZombieTexture so it's the same as in Fabric
-GuiAiManipulation|Forge: Changed the coordinates display of the selected entity to be only integers, just like in Fabric
-LoadstateMenu|Forge: Removed a sysout and added a crashfix
-IngameMenu|Forge: Enabled the loadstate button after savestating

+++ LoTAS-Forge/src/main/java/de/pfannekuchen/lotas/gui/GuiAiManipulation.java
@@ -115,7 +115,7 @@ public class GuiAiManipulation extends GuiScreen {
 		zText.drawTextBox();
 		
 		Vec entityPos = AIManipMod.getSelectedEntityPos();
-		drawCenteredString(MCVer.getFontRenderer(mc), AIManipMod.getSelectedEntity().getName() + " (" + entityPos.x + ", " + entityPos.y + ", " + entityPos.z + ")", width / 2, 5, 0xFFFFFF);
+		drawCenteredString(MCVer.getFontRenderer(mc), AIManipMod.getSelectedEntity().getName() + " (" + (int)entityPos.x + ", " + (int)entityPos.y + ", " + (int)entityPos.z + ")", width / 2, 5, 0xFFFFFF);
 	}
 	
 }
